Break up HeapObjects.cs, Rename DSasmDefs to simply Defs, Delete superfluous parameter descriptions
